+++ pkg/controller/firewall/controller.go
@@ -148,7 +148,7 @@ func hasFirewall(router *api.Router) bool {
 		return false
 	}
 	for _, resource := range router.Spec.Resources {
-		if resource.APIVersion == api.FirewallAPIVersion && (resource.Kind == "FirewallZone" || resource.Kind == "FirewallPolicy" || resource.Kind == "FirewallRule") {
+		if resource.APIVersion == api.FirewallAPIVersion && (resource.Kind == "FirewallZone" || resource.Kind == "FirewallPolicy" || resource.Kind == "FirewallRule" || resource.Kind == "ClientPolicy") {
 			return true
 		}
 	}