+++ pkg/render/nixos.go
@@ -1472,7 +1472,7 @@ func nixOSNeedsDnsmasq(router *api.Router) bool {
 func nixOSNeedsNftables(router *api.Router) bool {
 	for _, res := range router.Spec.Resources {
 		switch res.Kind {
-		case "IPv4SourceNAT", "NAT44Rule", "IPv4PolicyRoute", "IPv4PolicyRouteSet", "IPv4DefaultRoutePolicy", "FirewallZone", "FirewallPolicy", "FirewallRule", "FirewallLog", "PathMTUPolicy":
+		case "IPv4SourceNAT", "NAT44Rule", "IPv4PolicyRoute", "IPv4PolicyRouteSet", "IPv4DefaultRoutePolicy", "FirewallZone", "FirewallPolicy", "FirewallRule", "FirewallLog", "ClientPolicy", "PathMTUPolicy":
 			return true
 		}
 	}
@@ -1550,7 +1550,7 @@ func nixOSPackages(router *api.Router, host api.NixOSHostSpec) ([]string, []stri
 		case "DHCPv4Server", "DHCPv4Scope", "DHCPv6Server", "DHCPv6Scope":
 			service["dnsmasq"] = true
 			debug["dnsmasq"] = true
-		case "IPv4SourceNAT", "IPv4PolicyRoute", "IPv4PolicyRouteSet", "IPv4DefaultRoutePolicy", "FirewallZone", "FirewallPolicy", "FirewallRule", "PathMTUPolicy":
+		case "IPv4SourceNAT", "IPv4PolicyRoute", "IPv4PolicyRouteSet", "IPv4DefaultRoutePolicy", "FirewallZone", "FirewallPolicy", "FirewallRule", "ClientPolicy", "PathMTUPolicy":
 			service["nftables"] = true
 			debug["nftables"] = true
 		case "PPPoEInterface":